+++ simvue/api/objects/storage/s3.py
@@ -1,6 +1,4 @@
-"""
-Simvue S3 Storage
-=================
+"""Simvue S3 Storage.
 
 Class for interacting with an S3 based storage on the server.
 
@@ -9,28 +7,34 @@ Class for interacting with an S3 based storage on the server.
 import typing
 
 try:
-    from typing import Self
+    from typing import Self, override
 except ImportError:
-    from typing_extensions import Self
+    from typing_extensions import Self, override
 import pydantic
 
-from simvue.api.objects.base import write_only
+from simvue.api.objects.base import write_only, staging_check
 
-from .base import StorageBase, staging_check
+from .base import StorageBase
 from simvue.models import NAME_REGEX
 
 
 class S3Storage(StorageBase):
-    """
-    Simvue S3 Storage
-    ===================
+    """Simvue S3 Storage.
 
     This class is used to connect to/create S3 storage objects on the Simvue server,
     any modification of instance attributes is mirrored on the remote object.
 
     """
 
-    def __init__(self, identifier: str | None = None, **kwargs) -> None:
+    @override
+    def __init__(
+        self,
+        identifier: str | None = None,
+        *,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
+        **kwargs,
+    ) -> None:
         """Initialise a S3 Storage
 
         If an identifier is provided a connection will be made to the
@@ -41,19 +45,27 @@ class S3Storage(StorageBase):
         ----------
         identifier : str, optional
             the remote server unique id for the target folder
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
         **kwargs : dict
             any additional arguments to be passed to the object initialiser
         """
         self.config = Config(self)
-        super().__init__(identifier, **kwargs)
+        super().__init__(
+            identifier, server_url=server_url, server_token=server_token, **kwargs
+        )
         self._local_only_args += [
             "endpoint_url",
             "region_name",
             "access_key_id",
             "secret_access_key",
             "bucket",
+            "ca_cert",
         ]
 
+    @override
     @classmethod
     @pydantic.validate_call
     def new(
@@ -62,14 +74,17 @@ class S3Storage(StorageBase):
         name: typing.Annotated[str, pydantic.Field(pattern=NAME_REGEX)],
         disable_check: bool,
         endpoint_url: pydantic.HttpUrl,
-        region_name: str,
         access_key_id: str,
         secret_access_key: pydantic.SecretStr,
         bucket: str,
         is_tenant_useable: bool,
         is_default: bool,
         is_enabled: bool,
+        region_name: str | None = None,
+        ca_cert: pydantic.SecretStr | None = None,
         offline: bool = False,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
         **__,
     ) -> Self:
         """Create a new S3 storage object.
@@ -82,14 +97,16 @@ class S3Storage(StorageBase):
             whether to disable checks for this system
         endpoint_url : str
             endpoint defining the S3 upload URL
-        region_name : str
-            the region name associated with this storage system
         access_key_id : str
             the access key identifier for the storage
         secret_access_key : str
             the secret access key, stored as a secret string
         bucket : str
             the bucket associated with this storage system
+        region_name : str | None, optional
+            the region name associated with this storage system if applicable
+        ca_cert : str | None, optional
+            provide a CA certificate for this storage
         is_tenant_useable : bool
             whether this system is usable by the current tenant
         is_enabled : bool
@@ -98,6 +115,10 @@ class S3Storage(StorageBase):
             if this storage system should become the new is_default
         offline : bool, optional
             if this instance should be created in offline mode, is_default False
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
 
         Returns
         -------
@@ -107,12 +128,18 @@ class S3Storage(StorageBase):
         """
         _config: dict[str, str] = {
             "endpoint_url": endpoint_url.__str__(),
-            "region_name": region_name,
             "access_key_id": access_key_id,
             "secret_access_key": secret_access_key.get_secret_value(),
             "bucket": bucket,
         }
-        _storage = S3Storage(
+
+        if region_name:
+            _config["region_name"] = region_name
+
+        if ca_cert:
+            _config["ca_cert"] = ca_cert.get_secret_value()
+
+        _storage = cls(
             name=name,
             backend="S3",
             config=_config,
@@ -120,8 +147,10 @@ class S3Storage(StorageBase):
             is_tenant_useable=is_tenant_useable,
             is_default=is_default,
             is_enabled=is_enabled,
-            _read_only=False,
+            server_url=server_url,
+            server_token=server_token,
             _offline=offline,
+            _read_only=False,
         )
         _storage._staging |= _config
         return _storage
@@ -168,14 +197,9 @@ class Config:
 
     @property
     @staging_check
-    def region_name(self) -> str:
+    def region_name(self) -> str | None:
         """Retrieve the region name for this storage"""
-        try:
-            return self._sv_obj.get_config()["region_name"]
-        except KeyError as e:
-            raise RuntimeError(
-                "Expected key 'region_name' in alert definition retrieval"
-            ) from e
+        return self._sv_obj.get_config().get("region_name")
 
     @region_name.setter
     @write_only