+++ simvue/api/objects/tag.py
@@ -1,6 +1,4 @@
-"""
-Simvue Server Tag
-=================
+"""Simvue Server Tag.
 
 Contains a class for remotely connecting to a Simvue Tag, or defining
 a new tag given relevant arguments.
@@ -18,9 +16,9 @@ from simvue.models import DATETIME_FORMAT
 from collections.abc import Generator
 
 try:
-    from typing import Self
+    from typing import Self, override
 except ImportError:
-    from typing_extensions import Self
+    from typing_extensions import Self, override
 
 __all__ = ["Tag"]
 
@@ -35,16 +33,21 @@ class TagSort(Sort):
 
 
 class Tag(SimvueObject):
-    """
-    Simvue Tag
-    ==========
+    """Simvue Tag.
 
     This class is used to connect to/create tag objects on the Simvue server,
     any modification of instance attributes is mirrored on the remote object.
 
     """
 
-    def __init__(self, identifier: str | None = None, **kwargs) -> None:
+    @override
+    def __init__(
+        self,
+        identifier: str | None = None,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
+        **kwargs,
+    ) -> None:
         """Initialise a Tag
 
         If an identifier is provided a connection will be made to the
@@ -55,14 +58,29 @@ class Tag(SimvueObject):
         ----------
         identifier : str, optional
             the remote server unique id for the target folder
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
         **kwargs : dict
             any additional arguments to be passed to the object initialiser
         """
-        super().__init__(identifier, **kwargs)
+        super().__init__(
+            identifier, server_url=server_url, server_token=server_token, **kwargs
+        )
 
+    @override
     @classmethod
     @pydantic.validate_call
-    def new(cls, *, name: str, offline: bool = False, **kwargs) -> Self:
+    def new(
+        cls,
+        *,
+        name: str,
+        offline: bool = False,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
+        **kwargs,
+    ) -> Self:
         """Create a new Tag on the Simvue server.
 
         Parameters
@@ -71,6 +89,10 @@ class Tag(SimvueObject):
             name for the tag
         offline : bool, optional
             create this tag in offline mode, default False.
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
 
         Returns
         -------
@@ -78,7 +100,14 @@ class Tag(SimvueObject):
             tag object with staged attributes
         """
         _data: dict[str, typing.Any] = {"name": name}
-        return Tag(name=name, _read_only=False, _offline=offline, **kwargs)
+        return cls(
+            name=name,
+            server_url=server_url,
+            server_token=server_token,
+            _offline=offline,
+            _read_only=False,
+            **kwargs,
+        )
 
     @property
     @staging_check
@@ -131,6 +160,7 @@ class Tag(SimvueObject):
             else None
         )
 
+    @override
     @classmethod
     @pydantic.validate_call
     def get(
@@ -139,6 +169,8 @@ class Tag(SimvueObject):
         count: int | None = None,
         offset: int | None = None,
         sorting: list[TagSort] | None = None,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
         **kwargs,
     ) -> Generator[tuple[str, "SimvueObject"]]:
         """Get tags from the server.
@@ -151,6 +183,10 @@ class Tag(SimvueObject):
             start index for results, default is 0.
         sorting : list[dict] | None, optional
             list of sorting definitions in the form {'column': str, 'descending': bool}
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
 
         Yields
         ------
@@ -159,7 +195,7 @@ class Tag(SimvueObject):
             Tag object representing object on server
         """
         # There are currently no tag filters
-        kwargs.pop("filters", None)
+        _ = kwargs.pop("filters", None)
 
         _params: dict[str, str] = {}
 
@@ -169,6 +205,8 @@ class Tag(SimvueObject):
         return super().get(
             count=count,
             offset=offset,
+            server_url=server_url,
+            server_token=server_token,
             **_params,
             **kwargs,
         )