+++ simvue/run.py
@@ -1,6 +1,4 @@
-"""
-Simvue Run
-==========
+"""Simvue Run.
 
 Main class for recording metrics and information to Simvue during code execution.
 This forms the central API for users.
@@ -17,6 +15,8 @@ import warnings
 import humanfriendly
 import datetime
 import os
+from unyt import unyt_quantity
+from unyt.exceptions import UnitParseError
 
 import pydantic
 import re
@@ -52,6 +52,7 @@ from .models import (
     MetricKeyString,
     validate_timestamp,
     simvue_timestamp,
+    LogLevel,
 )
 from .system import get_system
 from .metadata import git_info, environment
@@ -183,6 +184,8 @@ class Run:
         self._executor = Executor(self)
         self._dispatcher: DispatcherBaseClass | None = None
 
+        self._meta_cache: dict[str, typing.Any] = {}
+
         self._folder: Folder | None = None
         self._term_color: bool = True
         self._grids: dict[str, str] = {}
@@ -282,9 +285,7 @@ class Run:
     ) -> None:
         logger.debug(
             "Automatically closing run '%s' in status %s",
-            self.id
-            if self._user_config.run.mode == "online" and self._sv_obj
-            else "unregistered",
+            self.id if self.mode == "online" and self._sv_obj else "unregistered",
             self._status,
         )
 
@@ -297,6 +298,11 @@ class Run:
     def duration(self) -> float:
         """Return current run duration"""
         return time.time() - self._start_time
+
+    @property
+    def mode(self) -> typing.Literal["offline", "online", "disabled"]:
+        """Return whether this run is offline."""
+        return self._user_config.run.mode
 
     @property
     def processes(self) -> list[psutil.Process]:
@@ -497,7 +503,9 @@ class Run:
             if category == "events":
                 _events = Events.new(
                     run=self.id,
-                    offline=self._user_config.run.mode == "offline",
+                    offline=self.mode == "offline",
+                    server_url=self._user_config.server.url,
+                    server_token=self._user_config.server.token,
                     events=buffer,
                 )
                 return _events.commit()
@@ -505,13 +513,17 @@ class Run:
                 _grid_metrics = GridMetrics.new(
                     run=self.id,
                     data=buffer,
-                    offline=self._user_config.run.mode == "offline",
+                    server_url=self._user_config.server.url,
+                    server_token=self._user_config.server.token,
+                    offline=self.mode == "offline",
                 )
                 return _grid_metrics.commit()
             else:
                 _metrics = Metrics.new(
                     run=self.id,
-                    offline=self._user_config.run.mode == "offline",
+                    offline=self.mode == "offline",
+                    server_url=self._user_config.server.url,
+                    server_token=self._user_config.server.token,
                     metrics=buffer,
                 )
                 return _metrics.commit()
@@ -541,7 +553,7 @@ class Run:
             if self._sv_obj.status != "running":
                 self._sv_obj.status = self._status
                 _changed = True
-            if self._user_config.run.mode == "offline":
+            if self.mode == "offline":
                 self._sv_obj.started = self._start_time
                 _changed = True
             if _changed:
@@ -711,14 +723,12 @@ class Run:
         self._term_color = not no_color
 
         self._folder = Folder.new(
-            path=folder, offline=self._user_config.run.mode == "offline"
+            path=folder,
+            offline=self.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         )
         self._folder.commit()  # type: ignore
-
-        if isinstance(visibility, str) and visibility not in ("public", "tenant"):
-            self._error(
-                "invalid visibility option, must be either None, 'public', 'tenant' or a list of users"
-            )
 
         if self._user_config.run.mode not in ("online", "offline"):
             self._error("invalid mode specified, must be online, offline or disabled")
@@ -735,7 +745,7 @@ class Run:
         if name and not re.match(r"^[a-zA-Z0-9\-\_\s\/\.:]+$", name):
             self._error("specified name is invalid")
             return False
-        elif not name and self._user_config.run.mode == "offline":
+        elif not name and self.mode == "offline":
             name = randomname.get_name()
 
         self._status = "running" if running else "created"
@@ -755,7 +765,10 @@ class Run:
         self._timer = time.time()
 
         self._sv_obj = RunObject.new(
-            folder=folder, offline=self._user_config.run.mode == "offline"
+            folder=folder,
+            offline=self._user_config.run.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         )
 
         if description:
@@ -1083,7 +1096,12 @@ class Run:
         """
         self._status = "running"
 
-        self._sv_obj = RunObject(identifier=run_id, _read_only=False)
+        self._sv_obj = RunObject(
+            identifier=run_id,
+            _read_only=False,
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
+        )
 
         self._sv_obj.status = self._status
         self._sv_obj.system = get_system()
@@ -1204,7 +1222,7 @@ class Run:
                         "Emissions metrics require resource metrics collection - make sure resource metrics are enabled!"
                     )
                     return False
-                if self._user_config.run.mode == "offline":
+                if self.mode == "offline":
                     # Create an emissions monitor with no API calls
                     self._emissions_monitor = CO2Monitor(
                         intensity_refresh_interval=None,
@@ -1356,9 +1374,11 @@ class Run:
     def log_event(
         self,
         message: str,
+        *,
         timestamp: typing.Annotated[
             datetime.datetime | str | None, pydantic.BeforeValidator(simvue_timestamp)
         ] = None,
+        log_level: LogLevel | None = None,
     ) -> bool:
         """Log event to the server
 
@@ -1369,6 +1389,9 @@ class Run:
         timestamp : datetime.datetime | str, optional
             manually specify the time stamp for this log, by default None
             if a string is provided, local time
+        log_level : str, optional
+            the logging level for this event, default is 'info',
+            requires server with version >=1.2.16
 
         Returns
         -------
@@ -1386,7 +1409,8 @@ class Run:
 
             run.log_event(
                 message="Good Night",
-                timestamp=datetime.datetime.now(datetime.UTC)
+                timestamp=datetime.datetime.now(datetime.UTC),
+                log_level="debug"
             )
         ```
         """
@@ -1405,7 +1429,23 @@ class Run:
             self._error("Cannot log events when not in the running state")
             return False
 
-        _data = {"message": message, "timestamp": timestamp}
+        # FIXME: Temporary, this will eventually be removed
+        import semver
+
+        _log_level_server_version = semver.parse("1.2.16")
+        if (
+            log_level
+            and self._user_config.run.mode != "offline"
+            and self._user_config.server_version < _log_level_server_version
+        ):
+            self._error("Log level is not supported on current server.")
+            return False
+
+        _data = {
+            "message": message,
+            "timestamp": timestamp,
+            "log_level": log_level or "info",
+        }
         self._dispatcher.add_item(
             _data, object_type="events", blocking=self._queue_blocking
         )
@@ -1590,7 +1630,9 @@ class Run:
                 name=grid_name,
                 grid=axes_ticks,
                 labels=axes_labels,
-                offline=self._user_config.run.mode == "offline",
+                offline=self.mode == "offline",
+                server_url=self._user_config.server.url,
+                server_token=self._user_config.server.token,
             )
             _new_grid.commit()
 
@@ -1611,7 +1653,9 @@ class Run:
         try:
             _grid_attach = Grid(
                 identifier=self._grids[grid_name]["id"],
-                offline=self._user_config.run.mode == "offline",
+                offline=self.mode == "offline",
+                server_url=self._user_config.server.url,
+                server_token=self._user_config.server.token,
             )
             _grid_attach.read_only(False)
             _grid_attach.attach_metric_for_run(self.id, metric_name)
@@ -1680,6 +1724,12 @@ class Run:
             )
         ```
         """
+
+        # If there are any metric units to be uploaded do so now
+        if _units := self._meta_cache.get("metrics"):
+            self.update_metadata({"simvue": {"metrics": _units}})
+            del self._meta_cache["metrics"]
+
         # TODO: When metrics and grids are combined into a single entity
         # this can be removed. For now need to separate tensor based metrics
         # from regular
@@ -1759,11 +1809,9 @@ class Run:
         self,
         obj: typing.Any,
         category: typing.Literal["input", "output", "code"],
-        name: typing.Optional[
-            typing.Annotated[str, pydantic.Field(pattern=NAME_REGEX)]
-        ] = None,
+        name: typing.Annotated[str, pydantic.Field(pattern=NAME_REGEX)] | None = None,
         allow_pickle: bool = False,
-        metadata: dict[str, typing.Any] = None,
+        metadata: dict[str, typing.Any] | None = None,
     ) -> bool:
         """Save an object to the Simvue server
 
@@ -1814,7 +1862,9 @@ class Run:
                 allow_pickling=allow_pickle,
                 storage=self._storage_id,
                 metadata=metadata,
-                offline=self._user_config.run.mode == "offline",
+                offline=self.mode == "offline",
+                server_url=self._user_config.server.url,
+                server_token=self._user_config.server.token,
             )
             _artifact.attach_to_run(self.id, category)
         except (ValueError, RuntimeError) as e:
@@ -1833,10 +1883,8 @@ class Run:
         file_type: str | None = None,
         preserve_path: bool = False,
         snapshot: bool = False,
-        name: typing.Optional[
-            typing.Annotated[str, pydantic.Field(pattern=NAME_REGEX)]
-        ] = None,
-        metadata: dict[str, typing.Any] = None,
+        name: typing.Annotated[str, pydantic.Field(pattern=NAME_REGEX)] | None = None,
+        metadata: dict[str, typing.Any] | None = None,
     ) -> bool:
         """Upload file to the server
 
@@ -1886,10 +1934,12 @@ class Run:
                 name=name or stored_file_name,
                 storage=self._storage_id,
                 file_path=file_path,
-                offline=self._user_config.run.mode == "offline",
+                offline=self.mode == "offline",
                 mime_type=file_type,
                 metadata=metadata,
                 snapshot=snapshot,
+                server_url=self._user_config.server.url,
+                server_token=self._user_config.server.token,
             )
             _artifact.attach_to_run(self.id, category)
         except (ValueError, RuntimeError) as e:
@@ -2049,11 +2099,7 @@ class Run:
             self._heartbeat_termination_trigger.set()
             self._heartbeat_thread.join()
 
-        if (
-            self._sv_obj
-            and self._user_config.run.mode == "offline"
-            and self._status != "created"
-        ):
+        if self._sv_obj and self.mode == "offline" and self._status != "created":
             self._user_config.offline.cache.joinpath(
                 "runs", f"{self.id}.closed"
             ).touch()
@@ -2184,13 +2230,17 @@ class Run:
         names = names or []
 
         if names and not ids:
-            if self._user_config.run.mode == "offline":
+            if self.mode == "offline":
                 self._error(
                     "Cannot retrieve alerts based on names in offline mode - please use IDs instead."
                 )
                 return False
             try:
-                if alerts := Alert.get(offline=self._user_config.run.mode == "offline"):
+                if alerts := Alert.get(
+                    offline=self.mode == "offline",
+                    server_url=self._user_config.server.url,
+                    server_token=self._user_config.server.token,
+                ):
                     ids += [id for id, alert in alerts if alert.name in names]
                 else:
                     self._error("No existing alerts")
@@ -2213,7 +2263,9 @@ class Run:
         """Check if an existing alert matches definition."""
         # If the alert already exists just add the existing one
         for _id, _existing_alert in Alert.get(
-            offline=self._user_config.run.mode == "offline"
+            offline=self.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         ):
             if _existing_alert == alert:
                 return _id
@@ -2295,7 +2347,9 @@ class Run:
             range_low=range_low,
             range_high=range_high,
             frequency=frequency or 60,
-            offline=self._user_config.run.mode == "offline",
+            offline=self.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         )
 
         # If the alert already exists just add the existing one
@@ -2383,7 +2437,9 @@ class Run:
             frequency=frequency,
             aggregation=aggregation,
             notification=notification,
-            offline=self._user_config.run.mode == "offline",
+            offline=self.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         )
 
         # If the alert already exists just add the existing one
@@ -2444,7 +2500,9 @@ class Run:
             pattern=pattern,
             notification=notification,
             frequency=frequency,
-            offline=self._user_config.run.mode == "offline",
+            offline=self.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         )
 
         # If the alert already exists just add the existing one
@@ -2502,7 +2560,9 @@ class Run:
             name=name,
             notification=notification,
             description=description,
-            offline=self._user_config.run.mode == "offline",
+            offline=self.mode == "offline",
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
         )
 
         # If the alert already exists just add the existing one
@@ -2552,7 +2612,7 @@ class Run:
             self._error("Please specify alert to update either by ID or by name.")
             return False
 
-        if name and self._user_config.run.mode == "offline":
+        if name and self.mode == "offline":
             self._error(
                 "Cannot retrieve alerts based on names in offline mode - please use IDs instead."
             )
@@ -2560,7 +2620,7 @@ class Run:
 
         if name:
             try:
-                if alerts := Alert.get(offline=self._user_config.run.mode == "offline"):
+                if alerts := Alert.get(offline=self.mode == "offline"):
                     identifier = next(
                         (id for id, alert in alerts if alert.name == name), None
                     )
@@ -2574,7 +2634,11 @@ class Run:
         if not identifier:
             self._error(f"Alert with name '{name}' could not be found.")
 
-        _alert = UserAlert(identifier=identifier)
+        _alert = UserAlert(
+            identifier=identifier,
+            server_url=self._user_config.server.url,
+            server_token=self._user_config.server.token,
+        )
         if not isinstance(_alert, UserAlert):
             self._error(
                 f"Cannot update state for alert '{identifier}' "
@@ -2586,3 +2650,40 @@ class Run:
         _alert.commit()
 
         return True
+
+    @prettify_pydantic
+    @pydantic.validate_call
+    def set_metric_units(
+        self,
+        metric_name: MetricKeyString,
+        *,
+        units: str,
+        mks_unit: str | None = None,
+        mks_conversion: float | None = None,
+    ) -> None:
+        """Define units for metrics.
+
+        Parameters
+        ----------
+        metric_name : str
+            name of metric to assign units to
+        units : str
+            unit symbol
+        label : str | None, optional
+            alternative longer name for units
+        """
+        self._meta_cache.setdefault("metrics", {})
+
+        try:
+            _unit_obj = unyt_quantity.from_string(units)
+            self._meta_cache["metrics"][metric_name] = {
+                "units": units,
+                "mks_conversion": mks_conversion or float(_unit_obj.in_mks().value),
+                "mks_units": mks_unit or f"{_unit_obj.in_mks().units}",
+            }
+        except UnitParseError:
+            self._meta_cache["metrics"][metric_name] = {
+                "units": units,
+                "mks_conversion": mks_conversion,
+                "mks_units": mks_unit,
+            }