+++ tests/cli/test_sender_command.py
@@ -1,24 +1,22 @@
 import tempfile
-from typing import Any
 import pytest
 import click.testing
-import os
+import pathlib
 
-from simvue.bin.sender import run
+from simvue.bin.sender import sender_cli
 from simvue.run import Run
 
-from conftest import create_test_run_offline, setup_test_run
+from conftest import setup_test_run
 
 @pytest.mark.cli
 def test_sender_command(request, monkeypatch) -> None:
     with tempfile.TemporaryDirectory() as tempd:
         monkeypatch.setenv("SIMVUE_OFFLINE_DIRECTORY", tempd)
-        _run = Run(mode="offline")
-        setup_test_run(_run, temp_dir=tempd, create_objects=True, request=request)
-        _run.close()
+        with Run(mode="offline") as run:
+            _test_run_data = setup_test_run(run, temp_dir=pathlib.Path(tempd), create_objects=True, request=request)
         _runner = click.testing.CliRunner()
         _result = _runner.invoke(
-            run,
+            sender_cli,
             [
                 "-i",
                 tempd