+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Change Log
 
+## [v2.5.4](https://github.com/simvue-io/python-api/releases/tag/v2.5.4) - 2026-05-28
+
+- Added `env` option to server profiles in configuration for environment variables.
+- Added fallback for S3 servers only supporting PUT.
+- Fix use of alternative serve specification.
+
+## [v2.5.3](https://github.com/simvue-io/python-api/releases/tag/v2.5.3) - 2026-05-15
+
+- Set Storage region to be optional
+
+## [v2.5.2](https://github.com/simvue-io/python-api/releases/tag/v2.5.2) - 2026-05-15
+
+- Fixed issue whereby providing alternative server URL or token arguments was ignored.
+- Fixed legacy event logging support.
+- Added `ca_cert` argument for S3 storages.
+
+## ~~[v2.5.1](https://github.com/simvue-io/python-api/releases/tag/v2.5.1) - 2026-05-06~~
+
+**Yanked from PyPi**
+
+- ~~Fixed legacy event logging support.~~
+
+## ~~[v2.5.1](https://github.com/simvue-io/python-api/releases/tag/v2.5.1) - 2026-05-06~~
+
+**Yanked from PyPi as issue not fixed.**
+
+- ~~Fixed legacy event logging support.~~
+
+## [v2.5.0](https://github.com/simvue-io/python-api/releases/tag/v2.5.0) - 2026-05-05
+
+- Added support for log level in events creation.
+- Added support for provision of units in metrics.
+- Fixed bug where token is not passed correctly on server version check.
+
 ## [v2.4.1](https://github.com/simvue-io/python-api/releases/tag/v2.4.1) - 2026-03-31
 
 - Moved to using `threading.Event` as termination trigger events and added deprecation notice for `multiprocessing.Event`.